requirements.txt
tqdm>=4.50.0
prettytable>=2.1.0
numpy>=1.18.0
Pillow>=8.0.0
matplotlib>=3.0.0
rasterio